Move is_font_file and FONT_EXTENSIONS to utils module
These are general-purpose utilities, not OS-specific config.
Tests moved alongside the code.
